requirements.txt
trivium-cli==0.1.0
